Create Split the number and name to calculate the average of every groups
Background: For text with names and numbers in the same column, there are four groups of abcd, each group has 6 parallels, and 45 times of data are recorded. Calculate the average of the 45 times of data in the four groups of abcd.

Solutions: First manually select the required 45 times of data, cbind() them into a table, then use a for loop to select the numbers, and average each row.